+++ core/src/main/kotlin/net/onedaybeard/graftt/graft/Annotations.kt
@@ -11,7 +11,7 @@ import org.objectweb.asm.tree.ClassNode
 import org.objectweb.asm.tree.FieldNode
 import org.objectweb.asm.tree.MethodNode
 
-private typealias TypeList = java.util.ArrayList<Type>
+internal typealias TypeList = java.util.ArrayList<Type>
 
 
 /** returns types declared in [Graft.Annotations.remove] or an empty set if not present */